♻️ refactor: collection 도메인 파일 이동 - /application/hooks/api/collection -> /api/collection
