--- src/application/hooks/api/collection/queryKey.ts
@@ -1,3 +1,0 @@
-export const QUERY_KEYS = {
-  getCollectionCheck: (memeId: number) => ["getCollectionCheck", memeId],
-} as const;
