+++ src/application/hooks/api/index.ts
@@ -1,3 +1,2 @@
 export * from "./account";
-export * from "./collection";
 export * from "./search";
